+++ src/Models/subscription/SubscriptionAddonPricingPlan.ts
@@ -1,0 +1,7 @@
+import SubscriptionAddonPricingPlanInterface from '../../Interfaces/subscription/SubscriptionAddonPricingPlan';
+
+export default class SubscriptionAddonPricingPlan {
+    constructor(parameters: SubscriptionAddonPricingPlanInterface) {
+        Object.assign(this, parameters);
+    }
+}